+++ semana9/labex/public/index.html
@@ -1,15 +1,11 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
   <head>
     <meta charset="utf-8" />
+    <link rel="icon" href="" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web site created using create-react-app"
-    />
-    
-    <title>labex</title>
+
+    <title>labeX</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
