+++ semana9/labex/src/App.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const App = () => {
+    return(
+     <>
+     APP JS
+     </>
+    )
+};
+
+export default App;
