+++ semana9/labex/src/pages/AdmLoginPage/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const AdmLoginPage = () => {
+    return(
+     <>
+     AdmLoginPage
+     </>
+    )
+};
+
+export default AdmLoginPage;
