+++ semana9/labex/src/pages/CandidateSignUpPage/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const CandidateSignUpPage = () => {
+    return(
+     <>
+     CandidateSignUpPage
+     </>
+    )
+};
+
+export default CandidateSignUpPage;
