+++ semana9/labex/src/pages/HomePage/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const HomePage = () => {
+    return(
+     <>
+     HomePage
+     </>
+    )
+};
+
+export default HomePage;
