+++ semana9/labex/src/pages/TripsPage/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const TripsPage = () => {
+    return(
+     <>
+     TripsPage
+     </>
+    )
+};
+
+export default TripsPage;
