--- semana9/projeto/labex/src/App.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const App = () => {
-  return(
-    <>
-    App
-    </>
-  )
-}
-
-export default App
