--- semana9/projeto/labex/src/components/ApplicationFormPage/ApplicationFormPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const ApplicationFormPage = () => {
-  return(
-    <>
-Para o usuário se candidatar à viagem, página com o form de inscrição.
-</>
-  )
-}
-
-export default ApplicationFormPage
