--- semana9/projeto/labex/src/components/CreateTripPage/CreateTripPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const CreateTripPage = () => {
-  return(
-    <>
-Para criarmos uma viagem
-</>
-  )
-}
-
-export default CreateTripPage
