--- semana9/projeto/labex/src/components/HomePage/HomePage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const HomePage = () => {
-  return(
-    <>
-    Para o usuário escolher entre login e formulário de inscrição
-    </>
-  )
-}
-
-export default HomePage
