--- semana9/projeto/labex/src/components/ListTripsPage/ListTripsPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const ListTripsPage = () => {
-  return(
-    <>
-Para vermos todas as viagens
-</>
-  )
-}
-
-export default ListTripsPage
