--- semana9/projeto/labex/src/components/LoginPage/LoginPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const LoginPage = () => {
-  return(
-    <>
- Para fazermos login como administrador
- </>
-  )
-}
-
-export default LoginPage
