--- semana9/projeto/labex/src/components/Router/Router.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Router = () => {
-  return(
-    <>
-Para administrar todas as rotas
-</>
-  )
-}
-
-export default Router
