--- semana9/projeto/labex/src/components/TripDetailsPage/TripDetailsPage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const TripDetailsPage = () => {
-  return(
-    <>
-Para criarmos uma viagem
-</>
-  )
-}
-
-export default TripDetailsPage
